feat: joystick functions updated
